requirements.txt
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
Django-Verify-Email==2.0.0
django-cors-headers==4.3.1